fix:replace match-case syntax with if-else chain to support lower versions of Python.

+++ termii/utils.py
@@ -9,26 +9,25 @@ from termii.schemas.shared import RequestData, RequestType
 
 def make_request(data: RequestData) -> Dict:
     """Handle requests dispatch."""
-    match data.type:
-        case RequestType.post:
-            response = requests.post(
-                url=data.url, json=data.payload if data.payload else {}
-            )
-            return response.json()
+    if data.type == RequestType.post:
+        response = requests.post(
+            url=data.url, json=data.payload if data.payload else {}
+        )
+        return response.json()
 
-        case RequestType.get:
-            response = requests.get(url=data.url)
-            return response.json()
+    elif data.type == RequestType.get:
+        response = requests.get(url=data.url)
+        return response.json()
 
-        case RequestType.patch:
-            response = requests.patch(
-                url=data.url, json=data.payload if data.payload else {}
-            )
-            return response.json()
+    elif data.type == RequestType.patch:
+        response = requests.patch(
+            url=data.url, json=data.payload if data.payload else {}
+        )
+        return response.json()
 
-        case RequestType.delete:
-            response = requests.delete(url=data.url)
-            return response.json()
+    elif data.type == RequestType.delete:
+        response = requests.delete(url=data.url)
+        return response.json()
 
 
 def get_random_string(count) -> str:
